+++ clients.txt
@@ -1,0 +1,11 @@
+Rui Manuel
+234987156
+4
+Rua Sem Fim / 200 / 5Esq / 1200‐001 / Lisboa
+10 ; 36 ; 2
+::::::::::
+Belmiro Miguel
+111987666
+2
+Avenida dos Grilos / 100 / ‐ / 2300‐101 / Coimbra
+100 ; -136 ; 20
